feat(bas): use Text from lib/ui
