+++ frontends/bmd/src/components/election_info.tsx
@@ -9,9 +9,8 @@ import {
 } from '@votingworks/types';
 import { formatLongDate } from '@votingworks/utils';
 
-import { Prose } from '@votingworks/ui';
+import { Prose, Text, NoWrap } from '@votingworks/ui';
 import { Seal } from './seal';
-import { Text, NoWrap } from './text';
 import { PrecinctSelection } from '../config/types';
 import { precinctSelectionName } from '../utils/precinct_selection';
