--- frontends/bmd/src/components/text.test.tsx
@@ -1,75 +1,0 @@
-import React from 'react';
-import { render, screen } from '@testing-library/react';
-
-import { Text, TextWithLineBreaks } from './text';
-
-it('outputs paragraph tag', () => {
-  const text = 'paragraph';
-  render(<Text>{text}</Text>);
-  const element = screen.getByText(text);
-  expect(element.tagName).toEqual('P');
-  expect(element).toMatchSnapshot();
-});
-
-it('outputs "span" tag specified by "as" prop', () => {
-  const text = 'Text in a span?';
-  render(<Text as="span">{text}</Text>);
-  const element = screen.getByText(text);
-  expect(element.tagName).toEqual('SPAN');
-  expect(element).toMatchSnapshot();
-});
-
-it('centers centered style', () => {
-  const { container } = render(<Text center>Centered Text?</Text>);
-  expect(container.firstChild).toMatchSnapshot();
-});
-
-it('displays muted style', () => {
-  const { container } = render(<Text muted>Muted Text?</Text>);
-  expect(container.firstChild).toMatchSnapshot();
-});
-
-it('displays error style', () => {
-  const { container } = render(<Text error>Error Text?</Text>);
-  expect(container.firstChild).toMatchSnapshot();
-});
-
-it('narrow style', () => {
-  const { container } = render(<Text narrow>Narrow Wrapper</Text>);
-  expect(container.firstChild).toMatchSnapshot();
-});
-
-it('renders italic text', () => {
-  const { container } = render(<Text italic>Italic text</Text>);
-  expect(container.firstChild).toMatchSnapshot();
-});
-
-it('converts line-breaks into <p> and <br/> tags', () => {
-  const { container } = render(
-    <TextWithLineBreaks
-      text={'I’m a paragraph.\n\nAnd I’m a paragraph with a\nline break.'}
-    />
-  );
-  expect(container).toMatchInlineSnapshot(`
-    <div>
-      <p
-        class=""
-      >
-        <span>
-          I’m a paragraph.
-        </span>
-      </p>
-      <p
-        class=""
-      >
-        <span>
-          And I’m a paragraph with a
-        </span>
-        <br />
-        <span>
-          line break.
-        </span>
-      </p>
-    </div>
-  `);
-});
