+++ frontends/bmd/src/pages/contest_page.tsx
@@ -2,7 +2,7 @@ import { assert } from '@votingworks/utils';
 import React, { useContext, useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { CandidateVote, OptionalYesNoVote } from '@votingworks/types';
-import { LinkButton, Screen, Prose } from '@votingworks/ui';
+import { LinkButton, Screen, Prose, Text } from '@votingworks/ui';
 
 import { ordinal } from '../utils/ordinal';
 
@@ -11,7 +11,6 @@ import { BallotContext } from '../contexts/ballot_context';
 import { CandidateContest } from '../components/candidate_contest';
 import { ElectionInfo } from '../components/election_info';
 import { Sidebar } from '../components/sidebar';
-import { Text } from '../components/text';
 import { YesNoContest } from '../components/yes_no_contest';
 import { SettingsTextSize } from '../components/settings_text_size';
 import { TextIcon } from '../components/text_icon';
