+++ frontends/election-manager/src/screens/manual_data_import_index_screen.tsx
@@ -3,7 +3,7 @@ import React, { ReactChild, useContext, useEffect, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import styled from 'styled-components';
 
-import { Table, TD, Prose } from '@votingworks/ui';
+import { Prose, Table, TD, Text } from '@votingworks/ui';
 import {
   ExternalTallySourceType,
   TallyCategory,
@@ -16,7 +16,6 @@ import { routerPaths } from '../router_paths';
 import { AppContext } from '../contexts/app_context';
 import { NavigationScreen } from '../components/navigation_screen';
 import { Button, SegmentedButton } from '../components/button';
-import { Text } from '../components/text';
 import {
   convertTalliesByPrecinctToFullExternalTally,
   getEmptyExternalTalliesByPrecinct,
