+++ frontends/election-manager/src/screens/overvote_combination_report_screen.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 import pluralize from 'pluralize';
 
 import { assert, format } from '@votingworks/utils';
-import { Table, TD, LogoMark, Prose } from '@votingworks/ui';
+import { LogoMark, Prose, Table, TD, Text } from '@votingworks/ui';
 import { ContestTallyMeta } from '@votingworks/types';
 import { NavigationScreen } from '../components/navigation_screen';
 import { PrintButton } from '../components/print_button';
@@ -14,7 +14,6 @@ import {
   getOvervotePairTallies,
   getContestTallyMeta,
 } from '../lib/votecounting';
-import { Text } from '../components/text';
 import { HorizontalRule } from '../components/horizontal_rule';
 
 const ContestMeta = styled.div`
